Separate out non-live organisation code
