+++ app/controllers/organisations_controller.rb
@@ -14,6 +14,7 @@ class OrganisationsController < ApplicationController
     @documents = Organisations::DocumentsPresenter.new(@organisation)
     @what_we_do = Organisations::WhatWeDoPresenter.new(@organisation)
     @people = Organisations::PeoplePresenter.new(@organisation)
+    @not_live = Organisations::NotLivePresenter.new(@organisation)
 
     respond_to do |f|
       f.html do
