+++ app/views/organisations/_separate_website.html.erb
@@ -2,7 +2,7 @@
 <%= render partial: 'header' %>
 
 <%= render "govuk_publishing_components/components/notice", {
-  title: @show.non_live_organisation_notice
+  title: @not_live.non_live_organisation_notice
 } %>
 
 <div class="grid-row organisations__margin-bottom">
